pyproject.toml
[project]
name = "rag-advance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.5",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
]
